extract table row into CompareTableRow
